requirements.txt
altgraph==0.10.2
appdirs==1.4.3
bdist-mpkg==0.5.0
bonjour-py==0.3
certifi==2020.4.5.1
click==7.1.1
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.0
enum34==1.1.10
filelock==3.0.12
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
futures==3.2.0
importlib-metadata==1.6.0
importlib-resources==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.3
macholib==1.5.1
MarkupSafe==1.1.1
matplotlib==1.3.1
modulegraph==0.10.4
mysql-connector==2.2.9
mysql-connector-python==8.0.11
numpy==1.8.0rc1
pathlib2==2.3.5
pipenv==2018.11.26
protobuf==3.11.3
py2app==0.7.3
PyMySQL==0.9.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
python-dotenv==0.13.0
pytz==2013.7
scandir==1.10.0
scipy==0.13.0b1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.16
typing==3.7.4.1
urllib3==1.24.1
virtualenv==20.0.18
virtualenv-clone==0.5.4
Werkzeug==1.0.1
xattr==0.6.4
zipp==1.2.0
zope.interface==4.1.1
